+++ src/app/accordion-block/accordion-block.component.ts
@@ -1,7 +1,8 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
 import { Store } from '@ngrx/store';
-import { StylesActions } from "../store/styles.actions";
+
+import { StylesActions } from '../store/styles.actions';
