+++ src/app/shared/interceptors/token.interceptor.ts
@@ -1,0 +1,22 @@
+import { Injectable } from '@angular/core';
+import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
+import { Observable } from 'rxjs';
+
+import { AuthService } from '../services/auth.service';
+
+@Injectable()
+export class TokenInterceptor implements HttpInterceptor{
+  constructor(private auth: AuthService) {
+  }
+
+  intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+    if (this.auth.showAuthStatus()) {
+      req = req.clone({
+        setHeaders: {
+          Authorization: this.auth.createToken(req)
+        }
+      })
+    }
+    return next.handle(req)
+  }
+}
