+++ src/app/shared/pipes/simple.pipe.ts
