+++ src/app/store/index.ts
@@ -1,5 +1,6 @@
 import { ActionReducerMap} from '@ngrx/store';
-import {ElementStyles} from "./interfaces";
+
+import {ElementStyles} from "../shared/interfaces/interfaces";
 import {ElementsStyleReducer} from "./styles.reducer";
