--- src/app/token.interceptor.ts
@@ -1,21 +1,0 @@
-import {Injectable} from "@angular/core";
-import {AuthService} from "./auth.service";
-import {HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from "@angular/common/http";
-import {Observable} from "rxjs";
-
-@Injectable()
-export class TokenInterceptor implements HttpInterceptor{
-  constructor(private auth: AuthService) {
-  }
-
-intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-  if(this.auth.showAuthStatus()){
-  req = req.clone({
-    setHeaders:{
-      Authorization: this.auth.createToken(req)
-    }
-  })
-}
-return next.handle(req)
-}
-}
